+++ src/main/java/blusunrize/immersiveengineering/mixin/coremods/client/LevelRendererMixin.java
@@ -1,0 +1,39 @@
+/*
+ * BluSunrize
+ * Copyright (c) 2023
+ *
+ * This code is licensed under "Blu's License of Common Sense"
+ * Details can be found in the license file in the root folder of this project
+ */
+
+package blusunrize.immersiveengineering.mixin.coremods.client;
+
+import blusunrize.immersiveengineering.client.utils.VertexBufferHolder;
+import com.mojang.blaze3d.vertex.PoseStack;
+import net.minecraft.client.Camera;
+import net.minecraft.client.renderer.GameRenderer;
+import net.minecraft.client.renderer.LevelRenderer;
+import net.minecraft.client.renderer.LightTexture;
+import org.joml.Matrix4f;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(LevelRenderer.class)
+public class LevelRendererMixin
+{
+	@Inject(method = "renderLevel", at = @At(
+			value = "INVOKE_STRING",
+			target = "Lnet/minecraft/util/profiling/ProfilerFiller;popPush(Ljava/lang/String;)V",
+			args = {"ldc=destroyProgress"}
+	))
+	public void afterTESRRender(
+			PoseStack matrixStackIn, float partialTicks, long finishTimeNano, boolean drawBlockOutline,
+			Camera activeRenderInfoIn, GameRenderer gameRendererIn, LightTexture lightmapIn,
+			Matrix4f projectionIn, CallbackInfo ci
+	)
+	{
+		VertexBufferHolder.afterTERRendering();
+	}
+}